+++ src/kernel/mem.h
@@ -20,7 +20,7 @@ struct page {
 #define MAX_SIZE_CLASS 35
 #define USABLE_PAGE_SIZE(block_size) (PAGE_SIZE - round_up((u64)sizeof(Page), (u64)block_size))
 
-static const u16 size_classes[MAX_SIZE_CLASS] = {8, 12, 16, 24, 32, 40, 48, 56, 64, 80, 96, 128, 160, 192, 216, 232, 256, 320, 352, 384, 448, 512, 768, 904, 1024, 1200, 1320, 1520, 1640, 1720, 2048};
+static const unsigned long long size_classes[MAX_SIZE_CLASS] = {8, 12, 16, 24, 32, 40, 48, 56, 64, 80, 96, 128, 160, 192, 216, 232, 256, 320, 352, 384, 448, 512, 768, 904, 1024, 1200, 1320, 1520, 1640, 1720, 2048};
 
 extern char end[];
 
@@ -28,8 +28,8 @@ typedef struct Page {
     struct Page* prev;
     struct Page* next;
     int free_list_num;
-    u16 block_size;
-    u16 free_list_offset;
+    unsigned long long block_size;
+    unsigned long long free_list_offset;
     bool in_free_list;
 } Page;
 
@@ -38,9 +38,9 @@ typedef struct {
     SpinLock locks[MAX_SIZE_CLASS];
 } PagedAllocator;
 
-void kinit_page(Page*, u16);
-u16 align_size(u16);
-int get_size_class(u16);
+void kinit_page(Page*, unsigned long long);
+unsigned long long align_size(unsigned long long);
+int get_size_class(unsigned long long);
 
 void kinit();
 u64 left_page_cnt();
@@ -48,7 +48,7 @@ u64 left_page_cnt();
 WARN_RESULT void *kalloc_page();
 void kfree_page(void *);
 
-WARN_RESULT void *kalloc(u16);
+WARN_RESULT void *kalloc(unsigned long long);
 void kfree(void *);
 
 WARN_RESULT void *get_zero_page();
